renames react package
