+++ apps/react-example-nested-stores/src/app/Manager.tsx
@@ -1,7 +1,6 @@
-import {createStoreValue} from "./store/value";
-import {store} from "@rx-store/react-rx-store";
-import {appRootEffect} from "./store/effects";
-
+import { createStoreValue } from './store/value';
+import { store } from '@rx-store/react';
+import { appRootEffect } from './store/effects';
 
 const value = createStoreValue();
 export const { Manager, context } = store(value, appRootEffect);
