+++ apps/react-example-nested-stores/src/app/index.tsx
@@ -1,14 +1,14 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./App";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
 import { Manager } from './Manager';
 
 ReactDOM.render(
   <React.StrictMode>
-    {/* Render the `react-rx-store` Provider with the context value & root effect */}
+    {/* Render the `react` Provider with the context value & root effect */}
     <Manager>
       <App />
     </Manager>
   </React.StrictMode>,
-  document.getElementById("root")
+  document.getElementById('root')
 );
