+++ apps/react-example-nested-stores/src/main.tsx
@@ -1,15 +1,14 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./app/App";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './app/App';
 import { Manager } from './app/Manager';
-
 
 ReactDOM.render(
   <React.StrictMode>
-    {/* Render the `react-rx-store` Provider with the context value & root effect */}
+    {/* Render the `react` Provider with the context value & root effect */}
     <Manager>
       <App />
     </Manager>
   </React.StrictMode>,
-  document.getElementById("root")
+  document.getElementById('root')
 );
