+++ apps/rx-store-website/docs/react/installation.md
@@ -1,18 +1,18 @@
 ---
 id: react-installation
-title: "Installation"
+title: 'Installation'
 ---
 
 ## NPM
 
 ```
-npm i @rx-store/react-rx-store @rx-store/core rxjs
+npm i @rx-store/react @rx-store/core rxjs
 ```
 
 ## Yarn
 
 ```
-yarn add @rx-store/react-rx-store @rx-store/core rxjs
+yarn add @rx-store/react @rx-store/core rxjs
 ```
 
 The `RX Store` library will add ~4kb minified to your app
