--- libs/react-rx-store/README.md
@@ -1,7 +1,0 @@
-# react-rx-store
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test react-rx-store` to execute the unit tests via [Jest](https://jestjs.io).
