--- libs/react-rx-store/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/react-rx-store';
