+++ libs/react/.babelrc
