+++ libs/react/.eslintrc
