+++ tsconfig.json
@@ -17,7 +17,7 @@
     "baseUrl": ".",
     "paths": {
       "@rx-store/core": ["libs/core/src/index.ts"],
-      "@rx-store/react-rx-store": ["libs/react-rx-store/src/index.ts"]
+      "@rx-store/react": ["libs/react/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
